requirements.txt
appdirs==1.4.2
cycler==0.10.0
decorator==4.0.11
imageio==2.1.2
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0
olefile==0.44
opencv-python==3.2.0.7
packaging==16.8
pandas==0.20.1
Pillow==4.0.0
pyparsing==2.1.10
PyQt5==5.8
python-dateutil==2.6.0
pytz==2016.10
PyWavelets==0.5.2
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
sip==4.19.1
six==1.10.0
sklearn==0.0
